+++ mathesar_ui/src/i18n/i18n-store.ts
@@ -1,0 +1,11 @@
+import type { Formatters, Locales, Translations } from './i18n-types';
+
+export const loadedLocales: Record<Locales, Translations> = {} as Record<
+  Locales,
+  Translations
+>;
+
+export const loadedFormatters: Record<Locales, Formatters> = {} as Record<
+  Locales,
+  Formatters
+>;
